(#1057) Merge NativeContractState into ContractState
